// src/interfaces/point.ts
export interface Point {
  //   altitudeLevels?: string;
  //   altitudeRestrictions?: string;
  //   cat1: string;
  //   cruiseFlightLevel?: string;
  fixDescription?: string;
  fixId?: string;
  //   formation?: string;
  //   iaf?: string;
  id: number;
  //   kcasMach?: string;
  latitude?: string;
  longitude?: string;
  //   lowLevel?: string;
  //   maxSpeed?: string;
  overfly?: '' | 'direct' | 'intercept';
  pointNumber?: string;
  //   slats?: string;
  //   speedType?: string;
  //   techSpeed?: string;
  //   time?: string;
  //   timeTolerance?: string;
  //   timeType?: string;
  turn?: '' | 'left' | 'right';
  turnType?: string;
}